requirements.txt
Flask==2.1.2
Pillow==9.1.1
PyMySQL==1.0.2